requirements.txt
# Core
python-dotenv
pydantic
cryptography
pyOpenSSL

# Azure
azure-identity
azure-keyvault-secrets
azure-mgmt-compute
azure-mgmt-resource
azure-mgmt-network

# AST / Analysis
astroid
libcst
networkx

# ML
transformers
torch
numpy
scikit-learn

# A2A / Messaging
msgpack
jsonschema

# mTLS / Networking
aiohttp

# Monitoring UI
fastapi
uvicorn[standard]
websockets
rich

# Security
cffi